Adapt to the monorepo
