+++ packages/buidler-autoextern/package.json
@@ -1,41 +1,41 @@
 {
-  "name": "@nomiclabs/buidler-autoexternal",
+  "name": "@nomiclabs/buidler-autoextern",
   "version": "0.0.1",
   "description": "Buidler plugin that generates testable contracts with automatically exported private methods",
-  "repository": "github:nomiclabs/buidler-autoexternal",
+  "repository": "github:nomiclabs/buidler",
+  "homepage": "https://github.com/nomiclabs/buidler/tree/master/packages/buidler-autoextern",
   "author": "Nomic Labs SRL",
   "license": "MIT",
-  "main": "dist/src/index.js",
-  "types": "dist/src/index.d.ts",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
+  "keywords": [
+    "ethereum",
+    "smart-contracts",
+    "buidler",
+    "buidler-plugin",
+    "truffle",
+    "testing"
+  ],
   "scripts": {
-    "lint:fix": "prettier --write 'src/**/*.{js,ts}' 'test/**/*.{js,ts}' && tslint --fix --config tslint.json --project tsconfig.json",
-    "lint": "tslint --config tslint.json --project tsconfig.json",
-    "test": "mocha",
-    "build": "tsc",
-    "buidl": "tsc",
-    "watch": "tsc -w"
+    "lint:fix": "../../node_modules/.bin/prettier --write 'src/**/*.{js,ts}' 'test/**/*.{js,ts}' && npm run lint-src -- --fix && npm run lint-tests -- --fix",
+    "lint": "npm run lint-src && npm run lint-tests",
+    "lint-tests": "../../node_modules/.bin/tslint --config tslint.json --project ./tsconfig.json",
+    "lint-src": "../../node_modules/.bin/tslint --config tslint.json --project src/tsconfig.json",
+    "test": "./scripts/run-tests.sh",
+    "build": "../../node_modules/.bin/tsc --build src",
+    "build-test": "../../node_modules/.bin/tsc --build .",
+    "clean": "rm -rf dist build-test"
   },
   "files": [
-    "dist/src/",
+    "dist/",
     "src/",
     "LICENSE",
     "README.md"
   ],
   "devDependencies": {
-    "@types/chai": "^4.1.7",
+    "@nomiclabs/buidler": "^1.0.0-beta.2",
     "@types/fs-extra": "^5.0.4",
-    "@types/mocha": "^5.2.5",
-    "@types/node": "^8.10.38",
-    "chai": "^4.2.0",
-    "dotenv": "^6.2.0",
-    "mocha": "^5.2.0",
-    "prettier": "^1.15.3",
-    "source-map-support": "^0.5.9",
-    "ts-node": "^8.0.2",
-    "tslint": "^5.12.0",
-    "tslint-config-prettier": "^1.17.0",
-    "tslint-plugin-prettier": "^2.0.1",
-    "typescript": "^3.2.2"
+    "chai": "^4.2.0"
   },
   "peerDependencies": {
     "@nomiclabs/buidler": "^1.0.0-beta.2"
