+++ packages/buidler-autoextern/scripts/run-tests.sh
@@ -1,0 +1,3 @@
+#!/usr/bin/env bash
+
+../../node_modules/.bin/mocha --exit
