+++ packages/buidler-autoextern/src/tsconfig.json
@@ -1,0 +1,15 @@
+{
+  "extends": "../../../config/typescript/tsconfig.json",
+  "compilerOptions": {
+    "outDir": "../dist",
+    "rootDirs": ["."],
+    "composite": true
+  },
+  "include": ["./**/*.ts"],
+  "exclude": [],
+  "references": [
+    {
+      "path": "../../buidler-core/src"
+    }
+  ]
+}
