+++ packages/buidler-autoextern/src/type-extensions.d.ts
@@ -1,0 +1,9 @@
+import "@nomiclabs/buidler/types";
+
+import { AutoexternConfig } from "./types";
+
+declare module "@nomiclabs/buidler/types" {
+  export interface BuidlerConfig {
+    autoextern?: Partial<AutoexternConfig>;
+  }
+}
